requirements.txt
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
certifi==2025.4.26
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.2
click==8.1.8
cloudinary==1.44.0
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.2
daphne==4.1.2
dj-database-url==2.3.0
Django==5.2
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-redis==5.4.0
django-storages==1.14.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
docopt==0.6.2
dotenv==0.9.9
drf-extra-fields==3.7.0
filetype==1.2.0
gunicorn==23.0.0
h11==0.16.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
loadenv==0.1.1
msgpack==1.1.0
packaging==25.0
pillow==11.2.1
pipreqs==0.4.13
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==25.0.0
python-dotenv==1.1.0
redis==6.0.0
requests==2.32.3
service-identity==24.2.0
setuptools==80.0.1
six==1.17.0
spark-parser==1.8.9
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
whitenoise==6.9.0
xdis==6.1.1
yarg==0.1.10
zope.interface==7.2
